+++ test/widget_test.dart
@@ -5,11 +5,9 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_app/main.dart';
-import 'package:flutter_app/firebase_options.dart';
 
 // Note: Testing Firebase apps often requires mocking.
 // Since I don't want to introduce complex mocking dependencies just for a basic check,